fix(sidebar): isolate collapse rerenders; compact session rows
Project group toggle lived in Shell and re-rendered the whole chat
Markdown stage. Move the list into ChatSidebar with its own store
subscription, debounce collapse persistence, and tighten row height
(avatar 20px, single-line title).
