+++ src/plugins/contributors/shell.tsx
@@ -1,36 +1,24 @@
-import { memo, useCallback, useEffect, useMemo, useState } from 'react'
+import { memo, useCallback, useEffect, useState } from 'react'
 import { Link, NavLink, useLocation, useNavigate } from 'react-router-dom'
 import type { Context } from 'cordis'
 import type { SlotProps } from '../registry/slots.ts'
 import { bindSnapshot, type Snapshot, type SnapshotService } from '../infrastructure/snapshot.ts'
-import { bindSessionView, type SessionListItem, type SessionViewService } from '../infrastructure/session-view.ts'
+import { bindSessionView, type SessionViewService } from '../infrastructure/session-view.ts'
 import { bindProjectView, type ProjectViewService } from '../infrastructure/project-view.ts'
 import { parseAppPath } from '../infrastructure/session-route.ts'
-import {
-  UNGROUPED_PROJECT_KEY,
-  groupSessionsByProject,
-} from '../infrastructure/session-groups.ts'
-import { useSidebarCollapseStore } from '../infrastructure/sidebar-collapse-store.ts'
 import {
   APP_MODULES,
   moduleIdFromPath,
   type AppModuleId,
 } from '../infrastructure/app-modules.ts'
 import { FishLogo } from './brand.tsx'
-import { SidebarMascot } from './mascot/sidebar-mascot.tsx'
-import { resolveSessionMascot } from './mascot/session-mascot.ts'
+import { ChatSidebar } from './chat-sidebar.tsx'
 import { FolderGlyph } from './chat/project-panel.tsx'
 import { DashboardModule } from './dashboard-module.tsx'
 import {
   LuBug,
-  LuChevronDown,
-  LuChevronRight,
-  LuLayoutDashboard,
   LuMessageSquare,
   LuPanelLeft,
-  LuPanelLeftClose,
-  LuPlus,
-  LuRadio,
 } from 'react-icons/lu'
 
 export const name = 'shell'
@@ -134,66 +122,46 @@ function WorkspaceModule() {
   )
 }
 
-const SessionRow = memo(function SessionRow({
-  item,
-  active,
-  busy,
+/** 主区与侧栏折叠态解耦：组折叠时不重跑 stage/composer。 */
+const AgentMainPanels = memo(function AgentMainPanels({
   view,
-  onDelete,
+  renderSlot,
 }: {
-  item: SessionListItem
-  active: boolean
-  busy: boolean
   view: string
-  onDelete: (item: SessionListItem) => void
+  renderSlot: SlotProps['renderSlot']
 }) {
-  const identity = resolveSessionMascot(item.id, item.mascot)
   return (
-    <div
-      className={`group mb-0.5 flex w-full items-stretch rounded-[8px] ${
-        active ? 'bg-[var(--dsw-business-soft)] text-[var(--dsw-business)]' : 'hover:bg-[var(--dsw-hover)]'
-      }`}
-    >
-      <Link
-        to={`/s/${item.id}${view === 'debug' ? '/debug' : ''}`}
-        className="flex min-w-0 flex-1 items-center gap-2.5 px-2 py-1.5 text-left text-sm"
+    <div className="relative flex min-h-0 w-full flex-1 flex-col overflow-hidden">
+      {/*
+        不用 display:none：大 Markdown DOM 被隐藏后再显示会重算布局（数百 ms longtask）。
+        absolute + visibility 保活布局，Chat↔Debug 只切可见性。
+      */}
+      <div
+        className={`absolute inset-0 min-h-0 overflow-hidden ${
+          view === 'chat' ? 'z-[1] flex' : 'pointer-events-none invisible z-0 flex'
+        }`}
+        aria-hidden={view !== 'chat'}
       >
-        <SidebarMascot
-          size={28}
-          sessionId={item.id}
-          identity={identity}
-          busy={busy}
-          title={`${identity.shape} · ${identity.color}`}
-        />
-        <span className="min-w-0 flex-1">
-          <div className="truncate font-medium">
-            {(item.type ?? 'chat') === 'live' ? (
-              <span className="mr-1.5 text-[10px] font-semibold tracking-wide text-[var(--dsw-label-3)] uppercase">
-                live
-              </span>
-            ) : null}
-            {item.title}
+        <div className="relative flex min-h-0 w-full flex-1 flex-col overflow-hidden">
+          <div className="flex min-h-0 flex-1 flex-col overflow-y-auto overscroll-contain px-6 py-4 pb-44 md:px-8 lg:px-10">
+            {renderSlot('stage')}
           </div>
-          <div className="mt-0.5 font-mono text-[10px] opacity-70">
-            {item.id.slice(0, 8)} · {item.eventCount} events
+          <div className="pointer-events-none absolute inset-x-0 bottom-0 z-[2] bg-transparent px-6 pb-4 md:px-8 lg:px-10">
+            <div className="pointer-events-auto space-y-2 bg-transparent">
+              {renderSlot('dock')}
+              {renderSlot('composer')}
+            </div>
           </div>
-        </span>
-      </Link>
-      <button
-        type="button"
-        className="shrink-0 px-2 text-[var(--dsw-label-3)] opacity-0 transition-opacity hover:text-[var(--dsw-danger)] group-hover:opacity-100 focus:opacity-100"
-        aria-label={`Delete session ${item.title}`}
-        title="Delete"
-        onClick={(event) => {
-          event.preventDefault()
-          event.stopPropagation()
-          onDelete(item)
-        }}
+        </div>
+      </div>
+      <div
+        className={`absolute inset-0 min-h-0 overflow-hidden ${
+          view === 'debug' ? 'z-[1] flex flex-col' : 'pointer-events-none invisible z-0 flex flex-col'
+        }`}
+        aria-hidden={view !== 'debug'}
       >
-        <svg viewBox="0 0 24 24" className="size-4" fill="none" stroke="currentColor" strokeWidth="1.8">
-          <path strokeLinecap="round" strokeLinejoin="round" d="M6 7h12M10 7V5h4v2m-6 3v8m4-8v8m-7-11 1 14h10l1-14" />
-        </svg>
-      </button>
+        {renderSlot('trajectory')}
+      </div>
     </div>
   )
 })
@@ -206,40 +174,19 @@ function Shell(props: SlotProps) {
   const navigate = useNavigate()
   const location = useLocation()
   const live = useSnapshot((state: Snapshot) => state.plugins.some((plugin) => plugin.enabled))
-  const sessions = useSessionView((state) => state.sessions)
   const sessionId = useSessionView((state) => state.sessionId)
   const project = useSessionView((state) => state.project)
   const view = useSessionView((state) => state.view)
-  const agentBusy = useSessionView((state) => state.agentStatus === 'running')
   const [settingsOpen, setSettingsOpen] = useState(false)
   const [sidebarCollapsed, setSidebarCollapsed] = useState(false)
-  const collapsedProjects = useSidebarCollapseStore((state) => state.collapsed)
-  const toggleProjectGroup = useSidebarCollapseStore((state) => state.toggle)
   const activeModule = moduleIdFromPath(location.pathname)
   const appRoute = parseAppPath(location.pathname)
   // 侧栏高亮跟 URL，不跟 store：点一下立刻亮，不等 load 完成
   const routeSessionId = appRoute.kind === 'session' ? appRoute.sessionId : null
   const agentHref = sessionId ? `/s/${sessionId}${view === 'debug' ? '/debug' : ''}` : '/'
   const showChatSidebar = activeModule === 'agent' && !sidebarCollapsed
-  const projectGroups = useMemo(() => groupSessionsByProject(sessions), [sessions])
-
-  function createChat(opts: { type?: 'chat' | 'live'; projectPath?: string } = {}) {
-    void sessionView.newSession(opts).then((id) => navigate(`/s/${id}`))
-  }
-
-  const deleteChat = useCallback(
-    (item: SessionListItem) => {
-      if (!window.confirm(`Delete session “${item.title}”?`)) return
-      const wasActive = item.id === sessionId
-      // 列表已乐观移除；切路由等 DELETE/load 完成后再跳，避免闪回
-      void sessionView.deleteSession(item.id).then(() => {
-        if (!wasActive) return
-        const next = sessionView.get().sessionId
-        navigate(next ? `/s/${next}` : '/')
-      })
-    },
-    [navigate, sessionId, sessionView],
-  )
+  const collapseSidebar = useCallback(() => setSidebarCollapsed(true), [])
+  const expandSidebar = useCallback(() => setSidebarCollapsed(false), [])
 
   // 单向：URL → sessionView。回写只靠 Link / navigate，不做 state→URL。
   useEffect(() => {
@@ -286,132 +233,13 @@ function Shell(props: SlotProps) {
         onSettings={() => setSettingsOpen(true)}
       />
 
-      {/* 模块切换用 hidden 保活，避免卸载重挂导致路由体感慢 */}
-      <aside
-        className={`app-side-bar min-h-0 flex-col overflow-hidden border-r border-[var(--dsw-border)] bg-[var(--dsw-sidebar)] ${
-          showChatSidebar ? 'flex' : 'hidden'
-        }`}
-        aria-hidden={!showChatSidebar}
-      >
-        <div className="min-h-0 flex-1 overflow-y-auto overscroll-contain px-2 pt-3 pb-3">
-          <div className="mb-1 flex items-center justify-between gap-2 px-2">
-            <span className="text-[11px] font-semibold tracking-wider text-[var(--dsw-label-3)] uppercase">Chat</span>
-            <button
-              type="button"
-              className="grid size-6 place-items-center rounded-[6px] text-[var(--dsw-label-3)] hover:bg-[var(--dsw-hover)] hover:text-[var(--dsw-business)]"
-              title="Collapse sidebar"
-              aria-label="Collapse sidebar"
-              onClick={() => setSidebarCollapsed(true)}
-            >
-              <LuPanelLeftClose className="size-3.5" />
-            </button>
-          </div>
-
-          <div className="app-side-actions" role="navigation" aria-label="Chat actions">
-            <button
-              type="button"
-              className="app-side-actions-item"
-              title="添加聊天"
-              aria-label="添加聊天"
-              onClick={() => createChat({ type: 'chat' })}
-            >
-              <span className="app-side-actions-icon" aria-hidden>
-                <LuPlus className="size-4" />
-              </span>
-              <span className="app-side-actions-label">添加聊天</span>
-            </button>
-            <button
-              type="button"
-              className="app-side-actions-item"
-              title="新建 Live"
-              aria-label="新建 Live"
-              onClick={() => createChat({ type: 'live' })}
-            >
-              <span className="app-side-actions-icon" aria-hidden>
-                <LuRadio className="size-4" />
-              </span>
-              <span className="app-side-actions-label">新建 Live</span>
-            </button>
-            <Link to="/dashboard" className="app-side-actions-item" title="Dashboard" aria-label="Dashboard">
-              <span className="app-side-actions-icon" aria-hidden>
-                <LuLayoutDashboard className="size-4" />
-              </span>
-              <span className="app-side-actions-label">Dashboard</span>
-            </Link>
-          </div>
-
-          <div className="mt-2 space-y-2">
-            {sessions.length === 0 ? (
-              <p className="px-2 text-[11px] leading-4 text-[var(--dsw-label-3)]">No chats yet. Send a message or create one.</p>
-            ) : (
-              projectGroups.map((group) => {
-                // 含当前路由会话的组首帧就展开，避免「先塌再展开」闪一下
-                const hasRouteSession = Boolean(
-                  routeSessionId && group.sessions.some((row) => row.id === routeSessionId),
-                )
-                const collapsed = Boolean(collapsedProjects[group.key]) && !hasRouteSession
-                const isUngrouped = group.key === UNGROUPED_PROJECT_KEY
-                return (
-                  <div key={group.key} className="min-w-0">
-                    <div className="group/header mb-0.5 flex items-center gap-0.5 px-1">
-                      <button
-                        type="button"
-                        className="flex min-w-0 flex-1 items-center gap-1.5 rounded-[6px] px-1 py-1 text-left text-[11px] font-semibold tracking-wide text-[var(--dsw-label-3)] uppercase hover:bg-[var(--dsw-hover)] hover:text-[var(--dsw-label)]"
-                        title={group.path ?? group.label}
-                        aria-expanded={!collapsed}
-                        onClick={() => toggleProjectGroup(group.key)}
-                      >
-                        {collapsed ? (
-                          <LuChevronRight className="size-3.5 shrink-0" />
-                        ) : (
-                          <LuChevronDown className="size-3.5 shrink-0" />
-                        )}
-                        {isUngrouped ? (
-                          <span className="grid size-3.5 place-items-center text-[10px] opacity-70" aria-hidden>
-                            —
-                          </span>
-                        ) : (
-                          <FolderGlyph className="size-3.5 shrink-0 opacity-80" />
-                        )}
-                        <span className="min-w-0 flex-1 truncate normal-case tracking-normal">{group.label}</span>
-                        <span className="shrink-0 font-mono text-[10px] opacity-60">{group.sessions.length}</span>
-                      </button>
-                      <button
-                        type="button"
-                        className="grid size-6 shrink-0 place-items-center rounded-[6px] text-[var(--dsw-label-3)] opacity-0 hover:bg-[var(--dsw-hover)] hover:text-[var(--dsw-business)] group-hover/header:opacity-100 focus:opacity-100"
-                        title={isUngrouped ? '在未分组下添加聊天' : `在 ${group.label} 下添加聊天`}
-                        aria-label={isUngrouped ? '在未分组下添加聊天' : `在 ${group.label} 下添加聊天`}
-                        onClick={() =>
-                          createChat({
-                            type: 'chat',
-                            ...(group.path ? { projectPath: group.path } : {}),
-                          })
-                        }
-                      >
-                        <LuPlus className="size-3.5" />
-                      </button>
-                    </div>
-                    {collapsed ? null : (
-                      <div className="min-w-0">
-                        {group.sessions.map((item) => (
-                          <SessionRow
-                            key={item.id}
-                            item={item}
-                            active={item.id === routeSessionId}
-                            busy={item.id === routeSessionId && agentBusy}
-                            view={view}
-                            onDelete={deleteChat}
-                          />
-                        ))}
-                      </div>
-                    )}
-                  </div>
-                )
-              })
-            )}
-          </div>
-        </div>
-      </aside>
+      <ChatSidebar
+        visible={showChatSidebar}
+        routeSessionId={routeSessionId}
+        useSessionView={useSessionView}
+        sessionView={sessionView}
+        onCollapse={collapseSidebar}
+      />
 
       <main className="flex min-h-0 min-w-0 flex-col overflow-hidden">
         <div
@@ -426,7 +254,7 @@ function Shell(props: SlotProps) {
                   className="chat-view-header-expand"
                   title="Expand sidebar"
                   aria-label="Expand sidebar"
-                  onClick={() => setSidebarCollapsed(false)}
+                  onClick={expandSidebar}
                 >
                   <LuPanelLeft className="size-3.5" />
                 </button>
@@ -469,38 +297,7 @@ function Shell(props: SlotProps) {
             </nav>
             <div className="chat-view-header-right" aria-hidden />
           </header>
-          <div className="relative flex min-h-0 w-full flex-1 flex-col overflow-hidden">
-            {/*
-              不用 display:none：大 Markdown DOM 被隐藏后再显示会重算布局（数百 ms longtask）。
-              absolute + visibility 保活布局，Chat↔Debug 只切可见性。
-            */}
-            <div
-              className={`absolute inset-0 min-h-0 overflow-hidden ${
-                view === 'chat' ? 'z-[1] flex' : 'pointer-events-none invisible z-0 flex'
-              }`}
-              aria-hidden={view !== 'chat'}
-            >
-              <div className="relative flex min-h-0 w-full flex-1 flex-col overflow-hidden">
-                <div className="flex min-h-0 flex-1 flex-col overflow-y-auto overscroll-contain px-6 py-4 pb-44 md:px-8 lg:px-10">
-                  {props.renderSlot('stage')}
-                </div>
-                <div className="pointer-events-none absolute inset-x-0 bottom-0 z-[2] bg-transparent px-6 pb-4 md:px-8 lg:px-10">
-                  <div className="pointer-events-auto space-y-2 bg-transparent">
-                    {props.renderSlot('dock')}
-                    {props.renderSlot('composer')}
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div
-              className={`absolute inset-0 min-h-0 overflow-hidden ${
-                view === 'debug' ? 'z-[1] flex flex-col' : 'pointer-events-none invisible z-0 flex flex-col'
-              }`}
-              aria-hidden={view !== 'debug'}
-            >
-              {props.renderSlot('trajectory')}
-            </div>
-          </div>
+          <AgentMainPanels view={view} renderSlot={props.renderSlot} />
         </div>
         <div
           className={activeModule === 'dashboard' ? 'flex min-h-0 flex-1 flex-col overflow-hidden' : 'hidden'}
