+++ src/style.css
@@ -310,9 +310,9 @@ body {
   align-items: center;
   gap: 8px;
   border-radius: 8px;
-  padding: 7px 8px;
+  padding: 5px 8px;
   color: var(--dsw-label-2);
-  font-size: 13px;
+  font-size: 12px;
   font-weight: 500;
   line-height: 1.2;
   text-align: left;
